images/task/requirements.txt
mysqlclient ~= 2.0.3
SQLAlchemy ~= 1.4.2
mysql-connector-python-rf
